refactor: make `iroh::rpc_protocol` private
