+++ iroh-cli/src/commands/console.rs
@@ -2,7 +2,7 @@ use anyhow::Result;
 use clap::{Parser, Subcommand};
 use colored::Colorize;
 use iroh::base::base32::fmt_short;
-use iroh::{client::Iroh, rpc_protocol::ProviderService};
+use iroh::client::{Iroh, ProviderService};
 use quic_rpc::ServiceConnection;
 use rustyline::{error::ReadlineError, Config, DefaultEditor};
 use tokio::sync::{mpsc, oneshot};
