+++ iroh/examples/client.rs
@@ -6,9 +6,7 @@
 //! run this example from the project root:
 //!     $ cargo run --example client
 use indicatif::HumanBytes;
-use iroh::{client::Entry, node::Node};
-use iroh_base::base32;
-use iroh_sync::store::Query;
+use iroh::{base::base32, client::docs::Entry, node::Node, sync::store::Query};
 use tokio_stream::StreamExt;
 
 #[tokio::main]
