+++ iroh/examples/collection-provide.rs
@@ -6,8 +6,7 @@
 //! This is using an in memory database and a random node id.
 //! run this example from the project root:
 //!     $ cargo run --example collection-provide
-use iroh::rpc_protocol::SetTagOption;
-use iroh_bytes::{format::collection::Collection, BlobFormat};
+use iroh::bytes::{format::collection::Collection, util::SetTagOption, BlobFormat};
 use tracing_subscriber::{prelude::*, EnvFilter};
 
 // set the RUST_LOG env var to one of {debug,info,warn} to see logging info
